Add files via upload
This update contains speed improvements, especially where it comes to processing the xml Rhythmbox databases.  Python class lists are now being used to cache the contents of the xml databases.  The first time the skill is activated, it will cache their contents.  After that, the skill will reference the cache.  The skill will need to be reloaded if there are changes made on the Rhythmbox side.