Various Enhancements Based on Forum Feedback
Based on feedback from Julie_Smith over on the Mycroft Community Forum, several enhancements were made to the code.  The ability to "refresh the Rhythmbox database" on request was added.  This will come in handy to individuals actively tweaking their Rhythmbox library and playlists while Mycroft is already active.  On top of that, major code revisions now rank probabilities as opposed to the old hard-coded pecking order.  Finally, along those lines, the code around keywords (such as "playlist", "album", "by", etc.) was tweaked.  Use keywords to enhance the ranking of requests where names are the same across options (such as a "country" playlist and "country" genre).